11 poc minimizer factory (#13)
* factory seems to be working

* seems to be working

* seems to be working

* code cleaning

* move code parts

* code cleaning

* pin pint dependency

* code comment

* seems to be working

* fitting_engine to minimizer_engine

* module renamed from Fitting to fitting

* force folder rename

* force folder rename

* more Fitting to fitting

* better naming

* code cleaning

* unit tests for fitter

* remove unused arguments

* unit tests

* more tests

* pin numpy to 1.26 after release of 2.0

* code comment and test fixes

* code cleaning

* typo
