+++ examples_old/example1_dream.py
@@ -19,7 +19,7 @@ def fit_fun(x):
 
 f = Fitter()
 f.initialize(b, fit_fun)
-f.switch_engine("bumps")
+f.switch_minimizer("bumps")
 
 x = np.array([1, 2, 3])
 y = np.array([2, 4, 6]) - 1
