+++ examples_old/example_dataset2pt2_broken.py
@@ -43,7 +43,7 @@ for idx, minimizer in enumerate(['lmfit', 'bumps', 'dfo_ls']):
 
     b.m = m_starting_point
     b.c = c_starting_point
-    f.switch_engine(minimizer)
+    f.switch_minimizer(minimizer)
 
     f_res = d['y'].easyscience.fit(f, vectorize=True)
     print(f_res.p)
