--- src/easyscience/Fitting/__init__.py
@@ -1,3 +1,0 @@
-from .fitter import Fitter  # noqa: F401, E402
-from .minimizers.minimizer_base import FitResults  # noqa: F401, E402
-from .multi_fitter import MultiFitter  # noqa: F401, E402
