+++ src/easyscience/Objects/Inferface.py
@@ -19,7 +19,7 @@ from typing import TypeVar
 _C = TypeVar('_C', bound=ABCMeta)
 _M = TypeVar('_M')
 if TYPE_CHECKING:
-    from easyscience.Fitting import Fitter
+    from easyscience.fitting import Fitter
 
 
 class InterfaceFactoryTemplate:
@@ -65,7 +65,7 @@ class InterfaceFactoryTemplate:
         :param new_interface: name of new interface to be created
         :type new_interface: str
         :param fitter: Fitting interface which contains the fitting object which may have bindings which will be updated.
-        :type fitter: EasyScience.Fitting.Fitting.Fitter
+        :type fitter: EasyScience.fitting.Fitter
         :return: None
         :rtype: noneType
         """
