+++ src/easyscience/fitting/__init__.py
@@ -1,0 +1,4 @@
+from .fitter import Fitter  # noqa: F401, E402
+from .minimizers.minimizer_base import FitResults  # noqa: F401, E402
+# Causes circular import
+# from .multi_fitter import MultiFitter  # noqa: F401, E402
