+++ src/easyscience/fitting/minimizers/__init__.py
@@ -1,0 +1,7 @@
+#  SPDX-FileCopyrightText: 2023 EasyScience contributors  <core@easyscience.software>
+#  SPDX-License-Identifier: BSD-3-Clause
+#  © 2021-2023 Contributors to the EasyScience project <https://github.com/easyScience/easyscience
+
+from .minimizer_base import MinimizerBase  # noqa: F401, E402
+from .utils import FitError  # noqa: F401, E402
+from .utils import FitResults  # noqa: F401, E402
