+++ tests/unit_tests/Fitting/test_constraints.py
@@ -10,8 +10,8 @@ from typing import Tuple
 
 import pytest
 
-from easyscience.Fitting.Constraints import NumericConstraint
-from easyscience.Fitting.Constraints import ObjConstraint
+from easyscience.fitting.Constraints import NumericConstraint
+from easyscience.fitting.Constraints import ObjConstraint
 from easyscience.Objects.Variable import Parameter
