+++ tests/unit_tests/Fitting/test_fitter.py
@@ -1,0 +1,251 @@
+from unittest.mock import MagicMock
+
+import pytest
+import numpy as np
+from easyscience.fitting.fitter import Fitter
+import easyscience.fitting.fitter
+
+
+class TestFitter():
+    @pytest.fixture
+    def fitter(self, monkeypatch):
+        monkeypatch.setattr(Fitter, '_update_minimizer', MagicMock())
+        self.mock_fit_object = MagicMock()
+        self.mock_fit_function = MagicMock()
+        return Fitter(self.mock_fit_object, self.mock_fit_function)
+
+    def test_constructor(self, fitter: Fitter):
+        # When Then Expect
+        assert fitter._fit_object == self.mock_fit_object
+        assert fitter._fit_function == self.mock_fit_function
+        assert fitter._dependent_dims is None
+        assert fitter._name_current_minimizer == 'lmfit-leastsq'
+        fitter._update_minimizer.assert_called_once_with('lmfit-leastsq')
+
+    def test_fit_constraints(self, fitter: Fitter):
+        # When
+        mock_minimizer = MagicMock()
+        mock_minimizer.fit_constraints = MagicMock(return_value='constraints')
+        fitter._minimizer = mock_minimizer
+
+        # Then
+        constraints = fitter.fit_constraints()
+
+        # Expect
+        assert constraints == 'constraints'
+
+    def test_add_fit_constraint(self, fitter: Fitter):
+        # When
+        mock_minimizer = MagicMock()
+        mock_minimizer.add_fit_constraint = MagicMock()
+        fitter._minimizer = mock_minimizer
+
+        # Then
+        fitter.add_fit_constraint('constraints')
+
+        # Expect
+        mock_minimizer.add_fit_constraint.assert_called_once_with('constraints')
+
+    def test_remove_fit_constraint(self, fitter: Fitter):
+        # When
+        mock_minimizer = MagicMock()
+        mock_minimizer.remove_fit_constraint = MagicMock()
+        fitter._minimizer = mock_minimizer
+
+        # Then
+        fitter.remove_fit_constraint(10)
+
+        # Expect
+        mock_minimizer.remove_fit_constraint.assert_called_once_with(10)
+
+    def test_make_model(self, fitter: Fitter):
+        # When
+        mock_minimizer = MagicMock()
+        mock_minimizer.make_model = MagicMock(return_value='model')
+        fitter._minimizer = mock_minimizer
+
+        # Then
+        model = fitter.make_model('pars')
+
+        # Expect
+        assert model == 'model'
+        mock_minimizer.make_model.assert_called_once_with('pars')
+
+    def test_evaluate(self, fitter: Fitter):
+        # When
+        mock_minimizer = MagicMock()
+        mock_minimizer.evaluate = MagicMock(return_value='result')
+        fitter._minimizer = mock_minimizer
+
+        # Then
+        result = fitter.evaluate('pars')
+
+        # Expect
+        assert result == 'result'
+        mock_minimizer.evaluate.assert_called_once_with('pars')
+
+    def test_convert_to_pars_obj(self, fitter: Fitter):
+        # When
+        mock_minimizer = MagicMock()
+        mock_minimizer.convert_to_pars_obj = MagicMock(return_value='obj')
+        fitter._minimizer = mock_minimizer
+
+        # Then
+        obj = fitter.convert_to_pars_obj('pars')
+
+        # Expect
+        assert obj == 'obj'
+        mock_minimizer.convert_to_pars_obj.assert_called_once_with('pars')
+
+    def test_initialize(self, fitter: Fitter):
+        # When
+        mock_fit_object = MagicMock()
+        mock_fit_function = MagicMock()
+
+        # Then
+        fitter.initialize(mock_fit_object, mock_fit_function)
+
+        # Expect
+        assert fitter._fit_object == mock_fit_object
+        assert fitter._fit_function == mock_fit_function
+        fitter._update_minimizer.count(2)
+
+    def test_create(self, fitter: Fitter):
+        # When
+        fitter._update_minimizer = MagicMock()
+
+        # Then
+        fitter.create('great-minimizer')
+
+        # Expect
+        fitter._update_minimizer.assert_called_once_with('great-minimizer')
+    
+    def test_switch_minimizer(self, fitter: Fitter):
+        # When
+        mock_minimizer = MagicMock()
+        mock_minimizer.fit_constraints = MagicMock(return_value='constraints')
+        mock_minimizer.set_fit_constraint = MagicMock()
+        fitter._minimizer = mock_minimizer
+
+        # Then
+        fitter.switch_minimizer('great-minimizer')
+
+        # Expect
+        fitter._update_minimizer.count(2)
+        mock_minimizer.set_fit_constraint.assert_called_once_with('constraints')
+        mock_minimizer.fit_constraints.assert_called_once()
+
+    def test_update_minimizer(self, monkeypatch):
+        # When
+        mock_fit_object = MagicMock()
+        mock_fit_function = MagicMock()
+
+        mock_string_to_enum = MagicMock(return_value=10)
+        mock_factory = MagicMock(return_value='minimizer')
+        monkeypatch.setattr(easyscience.fitting.fitter, 'from_string_to_enum', mock_string_to_enum)
+        monkeypatch.setattr(easyscience.fitting.fitter, 'factory', mock_factory)
+        fitter = Fitter(mock_fit_object, mock_fit_function)
+
+        # Then
+        fitter._update_minimizer('great-minimizer')
+
+        # Expect
+        assert fitter._name_current_minimizer == 'great-minimizer'
+        assert fitter._minimizer == 'minimizer'
+
+    def test_available_minimizers(self, fitter: Fitter):
+        # When
+        minimizers = fitter.available_minimizers
+
+        # Then Expect
+        assert minimizers == [
+            'LMFit', 'LMFit_leastsq', 'LMFit_powell', 'LMFit_cobyla',
+            'Bumps', 'Bumps_simplex', 'Bumps_newton', 'Bumps_lm',
+            'DFO', 'DFO_leastsq'
+        ]
+
+    def test_minimizer(self, fitter: Fitter):
+        # When
+        fitter._minimizer = 'minimizer'
+
+        # Then 
+        minimizer = fitter.minimizer
+
+        # Expect
+        assert minimizer == 'minimizer'
+
+    def test_fit_function(self, fitter: Fitter):
+        # When Then 
+        fit_function = fitter.fit_function
+
+        # Expect
+        assert fit_function == self.mock_fit_function
+
+    def test_set_fit_function(self, fitter: Fitter):
+        # When
+        fitter._name_current_minimizer = 'current_minimizer'
+
+        # Then
+        fitter.fit_function = 'new-fit-function'
+
+        # Expect
+        assert fitter._fit_function == 'new-fit-function'
+        fitter._update_minimizer.assert_called_with('current_minimizer')
+
+    def test_fit_object(self, fitter: Fitter):
+        # When Then 
+        fit_object = fitter.fit_object
+
+        # Expect
+        assert fit_object == self.mock_fit_object
+
+    def test_set_fit_object(self, fitter: Fitter):
+        # When
+        fitter._name_current_minimizer = 'current_minimizer'
+
+        # Then
+        fitter.fit_object = 'new-fit-object'
+
+        # Expect
+        assert fitter.fit_object == 'new-fit-object'
+        fitter._update_minimizer.assert_called_with('current_minimizer')
+
+    def test_fit(self, fitter: Fitter):
+        # When
+        fitter._precompute_reshaping = MagicMock(return_value=('x_fit', 'x_new', 'y_new', 'weights', 'dims'))
+        fitter._fit_function_wrapper = MagicMock(return_value='wrapped_fit_function')
+        fitter._post_compute_reshaping = MagicMock(return_value='fit_result')
+        fitter._minimizer = MagicMock()
+        fitter._minimizer.fit = MagicMock(return_value='result')
+
+        # Then
+        result = fitter.fit('x', 'y', 'weights', 'vectorized')
+
+        # Expect
+        fitter._precompute_reshaping.assert_called_once_with('x', 'y', 'weights', 'vectorized')
+        fitter._fit_function_wrapper.assert_called_once_with('x_new', flatten=True) 
+        fitter._post_compute_reshaping.assert_called_once_with('result', 'x', 'y')
+        assert result == 'fit_result'
+        assert fitter._dependent_dims == 'dims'
+        assert fitter._fit_function == self.mock_fit_function
+
+    def test_post_compute_reshaping(self, fitter: Fitter):
+        # When
+        fit_result = MagicMock()
+        fit_result.y_calc = np.array([[10], [20], [30]]) 
+        fit_result.y_err = np.array([[40], [50], [60]]) 
+        x = np.array([1, 2, 3])
+        y = np.array([4, 5, 6])
+
+        # Then
+        result = fitter._post_compute_reshaping(fit_result, x, y)
+
+        # Expect
+        assert np.array_equal(result.y_calc, np.array([10, 20, 30]))
+        assert np.array_equal(result.y_err, np.array([40, 50, 60])) 
+        assert np.array_equal(result.x, x)
+        assert np.array_equal(result.y_obs, y)
+
+# TODO
+#       def test_fit_function_wrapper()
+#       def test_precompute_reshaping()
