+++ tests/unit_tests/Objects/test_Groups.py
@@ -394,7 +394,7 @@ def test_baseCollection_constraints(cls):
     p1 = Parameter("p1", 1)
     p2 = Parameter("p2", 2)
 
-    from easyscience.Fitting.Constraints import ObjConstraint
+    from easyscience.fitting.Constraints import ObjConstraint
 
     p2.user_constraints["testing"] = ObjConstraint(p2, "2*", p1)
